Get rid of markdown conversion
Break dependency on `html-to-markdown`. That project inspired this, but
I realized it's not the same thing I want to do here, and also not efficient
enough. That project was started so I could grab wikipedia articles and store
them in text files on disk. This project is about reading in the browser without
distractions, and converting to and from markdown is not what is needed.

Also moved the app folder within the server folder because at some
point there will be a client (web browser) app distinct from the
server.