Get rid of double render needs
The ultimate goal of this is a plain content reader. With such a light weight
page there is no reason to send down the HTML content in JSON. That relic
was a holdover when I was sending down markdown, which might still have a purpose
but not in this app.

HTML is dropped in the page, JS code will do client interactions from now on.

Closes #16.