Use local temporary version of react view render
I intend to stop using react-view-render in the near future and use a non-lazy version of react components + redux in this project. It's good practice for me since I don't get to code as much anymore and I miss it tremendously.

This closes #18 and gets the project back into a working state.