Moving things around, also BORKED
App is currently borked on purpose. Components no longer modify window.location directly for page changes.

Moving things around until I'm happy. Not very happy with the link interceptor at the moment, think it should be some sort of decorator not a component since I have to pass in dispatch which is annoying.